Update a bit more, extract into a separate class
